Fixes for broken tests/warnings

+++ TestPatches/fileio.h
@@ -41,7 +41,7 @@ extern "C"{
         errx(1, "Data buffer not specified");
     if ((fd = creat(file_name, 0666)) < 1)
         errx(1, "Error creating file");
-    if (write(fd, data, size) < size)
+    if (write(fd, data, size) < ssize_t(size))
         errx(1, "Error writing samples");
     close(fd);
     return size;
